+++ src/router/index.tsx
@@ -2,6 +2,7 @@ import { createBrowserRouter, Navigate } from 'react-router-dom'
 import App from '@/app'
 import Login from '@/pages/login'
 import Home from '@/pages/home'
+import ComponentsDisplay from '@/pages/components-display'
 
 const router = createBrowserRouter([
   {
@@ -19,6 +20,10 @@ const router = createBrowserRouter([
       {
         path: 'home',
         element: <Home />,
+      },
+      {
+        path: 'components',
+        element: <ComponentsDisplay />,
       },
     ],
   },
